Adds better entrypoint execution
When you run another script inside the entrypoint script you have 2
shells running and when you tell docker-compose to stop the container
the inner program won't stop. So you have to force stop.

To avoid this we use exec instead so that we only have a single shell
running. This way stopping works as expected.